Bug after option update and refresh
After removing all options of a select and adding new ones, the move all button moves everything regardless of whether the user has filtered some of the options.  Since the move all method is just looking for filteredX data, the error was found inside of the filter method.  I am not sure why the additional find is performed, but just adding the data directly to the item appears to fix the issue.