Add: populate and sort function.
Allowed for me to manually set sort index of values and have them be
populated in the dual list box correctly. The goal behind this change was
so users can choose a particular series of elements from the list and post
them to the server, if there was an error then the model would be returned
and the chosen elements would end up back in the desired order if the
developer manually added the 'data-sortindex' attribute to the chosen
elements in the correct order, then called the populate and sort method.